refactor: Renamed Stream -> ActivityList. Created new component -> ActivityItem
This commit renames the Steam component to ActivityList because we handle activities which
is more clearer.

It also results in a new component called ActivityItem which represent one Activity. ActivityList
does not have the same responsibility as before because now it only renders the ActivityItem component. Before
it constructed and rendered the component.
